requirements.txt
requests
fastapi[all]
pydantic
beanie
motor
python-dotenv
email-validator
uvicorn
Jinja2